+++ R/teal_data-class.R
@@ -14,83 +14,103 @@ setOldClass("join_keys")
 #' @name teal_data-class
 #' @rdname teal_data-class
 #'
-#' @slot env (`environment`) environment containing data sets and possibly auxiliary variables.
-#'  Access variables with [get_var()] or [`[[`].
-#'  No setter provided. Evaluate code to add variables into `@env`.
-#' @slot code (`character`) vector representing code necessary to reproduce the contents of `@env`.
-#'  Access with [get_code()].
+#' @slot .xData (`environment`) environment containing data sets and possibly
+#'  auxiliary variables.
+#'  Access variables with [get()], [`$`], [teal.code::get_var()] or [`[[`].
+#'  No setter provided. Evaluate code to add variables into `@.xData`.
+#' @slot code (`list` of `character`) representing code necessary to reproduce the contents of `qenv`.
+#'  Access with [teal.code::get_code()].
 #'  No setter provided. Evaluate code to append code to the slot.
-#' @slot id (`integer`) random identifier assigned to each element of `@code`. Used internally.
-#' @slot warnings (`character`) vector of warnings raised when evaluating code.
-#'  Access with [get_warnings()].
-#' @slot messages (`character`) vector of messages raised when evaluating code.
-#' @slot join_keys (`join_keys`) object specifying joining keys for data sets in `@env`.
+#' @slot join_keys (`join_keys`) object specifying joining keys for data sets in
+#' `@.xData`.
 #'  Access or modify with [join_keys()].
-#' @slot datanames (`character`) vector of names of data sets in `@env`.
-#'  Used internally to distinguish them from auxiliary variables.
-#'  Access or modify with [datanames()].
-#' @slot verified (`logical(1)`) flag signifying that code in `@code` has been proven to yield contents of `@env`.
+#' @slot verified (`logical(1)`) flag signifying that code in `@code` has been
+#'  proven to yield contents of `@.xData`.
 #'  Used internally. See [`verify()`] for more details.
+#'
+#' @inheritSection teal.code::`qenv-class` Code
 #'
 #' @import teal.code
 #' @keywords internal
 setClass(
   Class = "teal_data",
   contains = "qenv",
-  slots = c(join_keys = "join_keys", datanames = "character", verified = "logical"),
+  slots = c(join_keys = "join_keys", verified = "logical"),
   prototype = list(
     join_keys = join_keys(),
-    datanames = character(0),
     verified = logical(0)
   )
 )
 
-#' Initialize `teal_data` object
+#' It initializes the `teal_data` class
 #'
-#' @name new_teal_data
+#' Accepts .xData as a list and converts it to an environment before initializing
+#' parent constructor (`qenv`).
+#' @noRd
+setMethod(
+  "initialize",
+  "teal_data",
+  function(.Object, .xData = list(), join_keys = join_keys(), code = list(), ...) { # nolint: object_name.
+    # Allow .xData to be a list and convert it to an environment
+    if (!missing(.xData) && inherits(.xData, "list")) {
+      .xData <- rlang::env_clone(list2env(.xData), parent = parent.env(.GlobalEnv)) # nolint: object_name.
+      lockEnvironment(.xData, bindings = TRUE)
+    }
+    args <- list(...)
+    checkmate::assert_environment(.xData)
+    checkmate::assert_class(join_keys, "join_keys")
+    checkmate::assert_list(args, names = "named")
+    if (!any(is.language(code), is.character(code))) {
+      stop("`code` must be a character or language object.")
+    }
+
+    if (is.language(code)) {
+      code <- paste(lang2calls(code), collapse = "\n")
+    }
+    if (length(code)) {
+      code <- paste(code, collapse = "\n")
+    }
+
+    methods::callNextMethod(
+      .Object,
+      .xData,
+      join_keys = join_keys,
+      verified = (length(args$code) == 0L && length(.xData) == 0L),
+      code = code2list(code),
+      ...
+    )
+  }
+)
+
+#' Reshape code to the list
 #'
-#' @param data (`named list`) of data objects.
-#' @param code (`character` or `language`) code to reproduce the `data`.
-#'   Accepts and stores comments also.
-#' @param join_keys (`join_keys`) object
-#' @param datanames (`character`) names of datasets passed to `data`.
-#'   Needed when non-dataset objects are needed in the `env` slot.
-#' @rdname new_teal_data
+#' List will be divided by the calls. Each element of the list contains `id` and `dependency` attributes.
+#'
+#' @param code `character` with the code.
+#'
+#' @return list of `character`s of the length equal to the number of calls in `code`.
+#'
 #' @keywords internal
-new_teal_data <- function(data,
-                          code = character(0),
-                          join_keys = join_keys(),
-                          datanames = union(names(data), names(join_keys))) {
-  checkmate::assert_list(data)
-  checkmate::assert_class(join_keys, "join_keys")
-  if (is.null(datanames)) datanames <- character(0) # todo: allow to specify
-  checkmate::assert_character(datanames)
-  if (!any(is.language(code), is.character(code))) {
-    stop("`code` must be a character or language object.")
+#' @noRd
+code2list <- function(code) {
+  checkmate::assert_character(code, null.ok = TRUE)
+  if (length(code) == 0) {
+    return(list())
   }
 
-  if (is.language(code)) {
-    code <- paste(lang2calls(code), collapse = "\n")
+  parsed_code <- parse(text = code, keep.source = TRUE)
+
+  code_list <- if (length(parsed_code)) {
+    lapply(split_code(code), function(current_code) {
+      parsed_code <- parse(text = current_code, keep.source = TRUE)
+      attr(current_code, "dependency") <- extract_dependency(parsed_code)
+      current_code
+    })
+  } else {
+    # empty code like "", or just comments
+    attr(code, "dependency") <- extract_dependency(parsed_code) # in case comment contains @linksto tag
+    list(code)
   }
-  if (length(code)) {
-    code <- paste(code, collapse = "\n")
-  }
-  verified <- (length(code) == 0L && length(data) == 0L)
-
-  id <- sample.int(.Machine$integer.max, size = length(code))
-
-  new_env <- rlang::env_clone(list2env(data), parent = parent.env(.GlobalEnv))
-  lockEnvironment(new_env, bindings = TRUE)
-
-  methods::new(
-    "teal_data",
-    env = new_env,
-    code = code,
-    warnings = rep("", length(code)),
-    messages = rep("", length(code)),
-    id = id,
-    join_keys = join_keys,
-    datanames = datanames,
-    verified = verified
-  )
+  names(code_list) <- sample.int(.Machine$integer.max, length(code_list))
+  code_list
 }