+++ inst/doc/teal-data-reproducibility.R
@@ -1,4 +1,5 @@
 ## ----message=FALSE, error=TRUE------------------------------------------------
+try({
 library(teal.data)
 
 data_empty <- teal_data()
@@ -10,7 +11,7 @@ data_with_data <- teal_data(i = head(iris), code = "i <- head(iris)")
 data_with_data # is unverified
 data_with_data <- within(data_with_data, i$rand <- sample(nrow(i)))
 data_with_data # remains unverified
-cat(get_code(data_with_data)) # warning is prepended
+})
 
 ## -----------------------------------------------------------------------------
 library(teal.data)
@@ -24,7 +25,7 @@ data_right <- teal_data(
     data <- data.frame(x = 11:20)
     data$id <- seq_len(nrow(data))
   })
-)
+) # is unverified
 (data_right_verified <- verify(data_right)) # returns verified object
 
 ## -----------------------------------------------------------------------------
@@ -36,5 +37,5 @@ data <- within(teal_data(), {
   head(i)
 })
 cat(get_code(data)) # retrieve all code
-cat(get_code(data, datanames = "i")) # retrieve code for `i`
+cat(get_code(data, names = "i")) # retrieve code for `i`